fix transition

+++ mobil.h
@@ -13,7 +13,7 @@ typedef struct
 	int sizeBoundary;
 } Mobil;
 
-Mobil makeMobil(Point pos, int w, int h);
+Mobil makeMobil(Point pos, double w, double h);
 void fillMobil(Frame f, Point p, int fillColor, int boundaryColor, int plus);
 void drawMobil(Mobil &m, Frame f, int color);
 void drawMobilFromPoints(Mobil &m, Point &p1, Point &p2, Point &p3, Point &p4, Frame &f, int color);
